Fix two warnings by the MSVC compiler
gtlsconnection-openssl.c(294): warning C4090: 'function': different 'const' qualifiers

tls\tests\dtls-connection.c(842): warning C4244: 'initializing': conversion from 'double' to 'gint64', possible loss of data

+++ tls/tests/dtls-connection.c
@@ -841,7 +841,7 @@ main (int   argc,
   };
   const TestData client_timeout = {
     0,  /* server_timeout */
-    0.5 * G_USEC_PER_SEC,  /* client_timeout */
+    (gint64) (0.5 * G_USEC_PER_SEC),  /* client_timeout */
     TRUE,  /* server_should_disappear */
     TRUE, /* server_should_close */
     G_TLS_AUTHENTICATION_NONE,  /* auth_mode */
